Move http request to api folder
