+++ src/stores/WalletStore.js
@@ -5,7 +5,7 @@ import { TransactionType, Token } from 'constants';
 import { Transaction, TransactionCost } from 'models';
 import { defineMessages } from 'react-intl';
 
-import axios from '../network/httpRequest';
+import axios from '../network/api';
 import Routes from '../network/routes';
 import { createTransferTx } from '../network/graphql/mutations';
 import { decimalToSatoshi } from '../helpers/utility';
